Adjust UNIMARC Schema to Avram 0.9.6 (#408)
* Adjust UNIMARC Schema

The following positions have been changed from `codes` to `flags` because they contain repeatable flags, filled with blanks:

- `100$a/17-19`
- `110$a/4-6`
- `120$a/3-6`, `120$a/9-16`
- `105$a/0-3`, `105$a/4-7`
- `115$a/11-14`
- `116$a/4-9`, `116$a/10-15`
- `117$a/2-7`
- `121$a/1-2`,
- `126$a/7-12`
- `140$a/0-3`, `140$a/4-7`, `140$a/9-16`
- `141$a/0-2`, `141$a/6-7`, `141$d/0-2`, `141$e/0-2`, `141$e/3-5`
- `145$b/5-6`, `145$c/5-6`, `145$d/5-6`
- `181$b/3-5`

In addition codes like "0-9", "1-9" must be made explicit as [0,]1,2,3,4,5,6,7,8,9:

- `145$b/7`, `145$c/7`, `145$d/7` and  `121$b/6`

In these cases the number of distinct codes is better be expressed with a pattern:

- `121$b/2-3`
- `135$a/5-9`

Additional documentation of the pattern requires <https://github.com/gbv/avram/issues/34> to be included in Avram specification.

* Adjust UNIMARC Schema with ISO 639-2 codelist

The codelists contained invalid ISO 639-2 codes and was repeated. Better
use a referenced codelist.